khelmitra_backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-rest-framework==0.1.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
Pillow==10.1.0